company added

+++ Tender.Models/Models/COMMON.cs
@@ -8,13 +8,13 @@ namespace Tender.Models.Models
 {
     public class COMMON
     {
-		public decimal IS_ACTIVE { get; set; }
+		public int IS_ACTIVE { get; set; }
 		public DateTime CREATE_DATE { get; set; }
 		public string CREATE_USER { get; set; }
 		public string CREATE_DEVICE { get; set; }
 		public DateTime UPDATE_DATE { get; set; }
 		public string UPDATE_USER { get; set; }
 		public string UPDATE_DEVICE { get; set; }
-		public decimal VER { get; set; }
+		public int VER { get; set; }
 	}
 }
